chore: apply code review - 2
